TASK: Change the storage to use two tables commit and stream
This change introduce a new table commit (the new source of truth), the
stream table is just used as an heuristic to help us implement snapshot or
other features. The commit table is updated in a single SQL request, so we
limit the change to have concurrency issue here. For commit that contains
a lots of events, that can also improve the performance when loading
content from the store (maybe)

+++ Classes/Command/EventStoreCommandController.php
@@ -35,6 +35,7 @@ class EventStoreCommandController extends CommandController
         $schema = $conn->getSchemaManager()->createSchema();
         $toSchema = clone $schema;
 
+        EventStoreSchema::createCommit($toSchema, $this->connectionFactory->getCommitName());
         EventStoreSchema::createStream($toSchema, $this->connectionFactory->getStreamName());
 
         $conn->beginTransaction();
@@ -58,7 +59,13 @@ class EventStoreCommandController extends CommandController
         $schema = $conn->getSchemaManager()->createSchema();
         $toSchema = clone $schema;
 
-        EventStoreSchema::drop($toSchema, $this->connectionFactory->getStreamName());
+        if ($schema->hasTable($this->connectionFactory->getCommitName())) {
+            EventStoreSchema::drop($toSchema, $this->connectionFactory->getCommitName());
+        }
+
+        if ($schema->hasTable($this->connectionFactory->getStreamName())) {
+            EventStoreSchema::drop($toSchema, $this->connectionFactory->getStreamName());
+        }
 
         $conn->beginTransaction();
         $statements = $schema->getMigrateToSql($toSchema, $conn->getDatabasePlatform());
